requirements.txt
flask
pytest
python-dateutil
pyserial
